externalize *add_decay_process* to util function
